# chronograf/server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chronograf",
    "description": "API endpoints for Chronograf",
    "version": "1.5.0.0"
  },
  "schemes": ["http"],
  "basePath": "/chronograf/v1",
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/sources": {
      "get": {
        "tags": ["sources"],
        "summary": "Configured data sources",
        "description": "These data sources store time series data.",
        "responses": {
          "200": {
            "description": "An array of data sources",
            "schema": {
              "$ref": "#/definitions/Sources"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["sources"],
        "summary": "Create new data source",
        "parameters": [
          {
            "name": "source",
            "in": "body",
            "description": "Configuration options for data source",
            "schema": {
              "$ref": "#/definitions/Source"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Data source successfully created",
            "headers": {
              "Location": {
                "type": "string",
                "format": "url",
                "description":
                  "Location of the newly created data source resource."
              }
            },
            "schema": {
              "$ref": "#/definitions/Source"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/sources/{id}": {
      "get": {
        "tags": ["sources"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          }
        ],
        "summary": "Configured data sources",
        "description": "These data sources store time series data.",
        "responses": {
          "200": {
            "description":
              "Data source used to supply time series information.",
            "schema": {
              "$ref": "#/definitions/Source"
            }
          },
          "404": {
            "description": "Unknown source id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "tags": ["sources"],
        "summary": "Update data source configuration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of a data source",
            "required": true
          },
          {
            "name": "config",
            "in": "body",
            "description": "data source configuration",
            "schema": {
              "$ref": "#/definitions/Source"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Data source's configuration was changed",
            "schema": {
              "$ref": "#/definitions/Source"
            }
          },
          "404": {
            "description":
              "Happens when trying to access a non-existent data source.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["sources"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          }
        ],
        "summary":
          "This specific data source will be removed from the data store.  All associated kapacitor resources and kapacitor rules resources are also removed.",
        "responses": {
          "204": {
            "description": "data source has been removed"
          },
          "404": {
            "description": "Unknown data source id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/sources/{id}/users": {
      "get": {
        "tags": ["sources", "users"],
        "summary": "Retrieve all data sources users",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Listing of all users",
            "schema": {
              "$ref": "#/definitions/InfluxDB-Users"
            }
          },
          "404": {
            "description": "Data source id does not exist.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["sources", "users"],
        "summary": "Create new user for this data source",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          },
          {
            "name": "user",
            "in": "body",
            "description": "Configuration options for new user",
            "schema": {
              "$ref": "#/definitions/InfluxDB-User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully created",
            "headers": {
              "Location": {
                "type": "string",
                "format": "url",
                "description": "Location of the newly created user resource."
              }
            },
            "schema": {
              "$ref": "#/definitions/InfluxDB-User"
            }
          },
          "404": {
            "description": "Data source id does not exist.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/sources/{id}/users/{user_id}": {
      "get": {
        "tags": ["sources", "users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          },
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "description": "ID of the specific user",
            "required": true
          }
        ],
        "summary": "Returns information about a specific user",
        "description": "Specific User within a data source",
        "responses": {
          "200": {
            "description": "Information relating to the user",
            "schema": {
              "$ref": "#/definitions/InfluxDB-User"
            }
          },
          "404": {
            "description": "Unknown user or unknown source",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "tags": ["sources", "users"],
        "summary": "Update user configuration",
        "description":
          "Update one parameter at a time (one of password, permissions or roles)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          },
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "description": "ID of the specific user",
            "required": true
          },
          {
            "name": "config",
            "in": "body",
            "description": "user configuration",
            "schema": {
              "$ref": "#/definitions/InfluxDB-User"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Users's configuration was changed",
            "schema": {
              "$ref": "#/definitions/InfluxDB-User"
            }
          },
          "404": {
            "description":
              "Happens when trying to access a non-existent user or source.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["sources", "users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the data source",
            "required": true
          },
          {
            "name": "user_id",
            "in": "path",
            "type": "string",
            "description": "ID of the specific user",
            "required": true
          }
        ],
        "summary": "This specific user will be removed from the data source",
        "responses": {
          "204": {
            "description": "User has been removed"
          },
          "404": {
            "description": "Unknown user id or data source",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/sources/{id}/services": {
      "get": {
        "tags": ["sources", "services"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          }
        ],
        "summary": "Retrieve list of services for a source",
        "responses": {
          "200": {
            "description": "An array of services",
            "schema": {
              "$ref": "#/definitions/Services"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["sources", "services"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "service",
            "in": "body",
            "description": "Configuration options for the service",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          }
        ],
        "summary": "Create a new service",
        "responses": {
          "200": {
            "description": "Returns the newly created service",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          },
          "504": {
            "description": "Gateway timeout happens when the server cannot connect to the service",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/sources/{id}/services/{srv_id}": {
      "get": {
        "tags": ["sources", "services"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "srv_id",
            "in": "path",
            "type": "string",
            "description": "ID of the service",
            "required": true
          }
        ],
        "summary": "Retrieve a service",
        "description": "Retrieve a single service by id",
        "responses": {
          "200": {
            "description": "Service connection information",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          },
          "404": {
            "description": "Unknown data source or service id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "tags": ["sources", "services"],
        "summary": "Update service configuration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "srv_id",
            "in": "path",
            "type": "string",
            "description": "ID of a service backend",
            "required": true
          },
          {
            "name": "service",
            "in": "body",
            "description": "service configuration",
            "schema": {
              "$ref": "#/definitions/Service"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Service configuration was changed",
            "schema": {
              "$ref": "#/definitions/Service"
            }
          },
          "504": {
            "description": "Gateway timeout happens when the server cannot connect to the service",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "422": {
            "description": "Unprocessable entity happens when the service ID provided does not exist",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "A processing or an unexpected error.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["sources", "services"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the source",
            "required": true
          },
          {
            "name": "srv_id",
            "in": "path",
            "type": "string",
            "description": "ID of the service",
            "required": true
          }
        ],
        "summary": "Remove Service backend",
        "description":
          "This specific service will be removed.",
        "responses": {
          "204": {
            "description": "service has been removed."
          },
          "404": {
            "description": "Unknown Data source or Service id",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Unexpected internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["organizations", "users"],
        "summary":
          "Retrieve all Chronograf users within the current organization",
        "description":
          "Returns all Chronograf users within the current organization from the store",
        "responses": {
          "200": {
            "description": "Successfully retrieved all users from the store",
            "schema": {
              "$ref": "#/definitions/Users"
            }
          },
          "400": {
            "description": "Failed to load users from store",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden to access this route",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": ["organizations", "users"],
        "summary": "Create new user",
        "description": "Creates a Chronograf user in the store",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully created",
            "headers": {
              "Location": {
                "type": "string",
                "format": "url",
                "description": "Location of the newly created user resource"
              }
            },
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description":
              "Invalid JSON – unable to encode or decode; or failed to perform operation in data store",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized to perform this operation",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden to access this route",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "422": {
            "description": "Invalid data schema provided to server for user",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["organizations", "users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the user",
            "required": true
          }
        ],
        "summary": "Retrieve a specific user",
        "description": "Returns a specific user from the store",
        "responses": {
          "200": {
            "description": "An User object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description":
              "Failed to load user from store; or failed to parse user ID as valid",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden to access this route",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "tags": ["organizations", "users"],
        "summary": "Update existing user",
        "description": "Updates a Chronograf user in the store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the user",
            "required": true
          },
          {
            "name": "user",
            "in": "body",
            "description": "Updated user",
            "schema": {
              "$ref": "#/definitions/User"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully updated",
            "headers": {
              "Location": {
                "type": "string",
                "format": "url",
                "description": "Location of the updated user resource"
              }
            },
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description":
              "Invalid JSON – unable to encode or decode; failed to parse user id as valid; or failed to perform operation in data store",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized to perform operation",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden to access this route",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "422": {
            "description": "Invalid data schema provided to server for user",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": ["organizations", "users"],
        "summary": "Delete user",
        "description": "Deletes a Chronograf user in the store",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "description": "ID of the user",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted"
          },
          "400": {
            "description":
              "Failed to parse user id as valid; failed to retrieve user from context; or failed to perform operation in data store",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "403": {
            "description": "Forbidden to access this route",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "default": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Users": {
      "type": "object",
      "required": ["users"],
      "properties": {
        "links": {
          "type": "object",
          "properties": {
            "self": {
              "type": "string",
              "description": "Self link mapping to this resource",
              "format": "url"
            }
          }
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
...
    "Services": {
      "type": "object",
      "required": ["services"],
      "properties": {
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Service"
          }
        }
      }
    },
...
    "Sources": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Source"
      }
    },
...
    "InfluxDB-Users": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/InfluxDB-User"
          }
        }
      },
      "example": {
        "users": [
          {
            "name": "docbrown",
            "permissions": [
              {
                "scope": "all",
                "allowed": [
                  "ViewAdmin",
                  "ViewChronograf",
                  "CreateDatabase",
                  "CreateUserAndRole",
                  "DropDatabase",
                  "DropData",
                  "ReadData",
                  "WriteData",
                  "ManageShard",
                  "ManageContinuousQuery",
                  "ManageQuery",
                  "ManageSubscription",
                  "Monitor",
                  "KapacitorAPI"
                ]
              }
            ],
            "roles": [
              {
                "name": "timetravelers",
                "permissions": [
                  {
                    "scope": "database",
                    "name": "telegraf",
                    "allowed": ["ReadData", "WriteData"]
                  }
                ],
                "links": {
                  "self": "/chronograf/v1/sources/3/roles/timetravelers"
                }
              }
            ],
            "links": {
              "self": "/chronograf/v1/sources/3/users/docbrown"
            }
          }
        ]
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}